Remove code for executable, rename folder to comit_rs_settings
