+++ src/comit_rs_settings/cnd.rs
@@ -1,4 +1,3 @@
-use crate::executable::Program;
 use rand;
 use rand::Rng;
 use serde::Serialize;
@@ -84,11 +83,4 @@ impl Default for Btsieve {
             },
         }
     }
-}
-
-pub struct Cnd;
-
-impl Program for Cnd {
-    const COMMAND: &'static str = "cnd";
-    const LOG_READY: &'static str = "Starting HTTP server on";
 }
