+++ src/lib.rs
@@ -1,5 +1,5 @@
+pub mod comit_rs_settings;
 pub mod create_comit_app;
 pub mod docker;
-pub mod executable;
 pub mod new;
 pub mod start_env;
